chore: fix typescript issue
