+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export { default as breakpoints, Breakpoint, Breakpoints } from './breakpoints';
+export { default as breakpoints } from './breakpoints';
 export { default as colors } from './colors';
 export { default as fonts } from './fonts';
 export { default as spacing } from './spacing';
